address comments

+++ src/ioda-stats/calcstats.cc
@@ -210,7 +210,7 @@ std::vector<int> update_mask(std::vector<float> maskvalues,
                              float maxval, 
                              const std::vector<int>& inputMask) {
   std::vector<int> updatedMask = inputMask;
-  for (int i = 0; i < maskvalues.size(); i++) {
+  for (size_t i = 0; i < maskvalues.size(); i++) {
     if (maskvalues[i] < minval || maskvalues[i] >= maxval) {
       updatedMask[i] = 1;
     }
